Clean up prior tests
